update locale switcher
